artisty-frontend/backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai>=1.0.0